+++ packages/pangea-sdk/pangea/services/audit/audit.py
@@ -17,6 +17,7 @@ from pangea.services.audit.models import (
     Event,
     EventEnvelope,
     EventVerification,
+    ExportRequest,
     LogBulkRequest,
     LogBulkResult,
     LogEvent,
@@ -750,6 +751,67 @@ class Audit(ServiceBase, AuditBase):
             self.update_published_roots(response.result)
         return self.handle_results_response(response, verify_consistency, verify_events)
 
+    def export(
+        self,
+        *,
+        format: DownloadFormat = DownloadFormat.CSV,
+        start: Optional[datetime.datetime] = None,
+        end: Optional[datetime.datetime] = None,
+        order: Optional[SearchOrder] = None,
+        order_by: Optional[str] = None,
+        verbose: bool = True,
+    ) -> PangeaResponse[PangeaResponseResult]:
+        """
+        Export from the audit log
+
+        Bulk export of data from the Secure Audit Log, with optional filtering.
+
+        OperationId: audit_post_v1_export
+
+        Args:
+            format: Format for the records.
+            start: The start of the time range to perform the search on.
+            end: The end of the time range to perform the search on. If omitted,
+                then all records up to the latest will be searched.
+            order: Specify the sort order of the response.
+            order_by: Name of column to sort the results by.
+            verbose: Whether or not to include the root hash of the tree and the
+                membership proof for each record.
+
+        Raises:
+            AuditException: If an audit based api exception happens
+            PangeaAPIException: If an API Error happens
+
+        Examples:
+            export_res = audit.export(verbose=False)
+
+            # Export may take several dozens of minutes, so polling for the result
+            # should be done in a loop. That is omitted here for brevity's sake.
+            try:
+                audit.poll_result(request_id=export_res.request_id)
+            except AcceptedRequestException:
+                # Retry later.
+
+            # Download the result when it's ready.
+            download_res = audit.download_results(request_id=export_res.request_id)
+            download_res.result.dest_url
+            # => https://pangea-runtime.s3.amazonaws.com/audit/xxxxx/search_results_[...]
+        """
+        input = ExportRequest(
+            format=format,
+            start=start,
+            end=end,
+            order=order,
+            order_by=order_by,
+            verbose=verbose,
+        )
+        try:
+            return self.request.post(
+                "v1/export", PangeaResponseResult, data=input.dict(exclude_none=True), poll_result=False
+            )
+        except pexc.AcceptedRequestException as e:
+            return e.response
+
     def log_stream(self, data: dict) -> PangeaResponse[PangeaResponseResult]:
         """
         Log streaming endpoint
@@ -815,7 +877,10 @@ class Audit(ServiceBase, AuditBase):
         return self.request.post("v1/root", RootResult, data=input.dict(exclude_none=True))
 
     def download_results(
-        self, result_id: str, format: Optional[DownloadFormat] = None
+        self,
+        result_id: Optional[str] = None,
+        format: DownloadFormat = DownloadFormat.CSV,
+        request_id: Optional[str] = None,
     ) -> PangeaResponse[DownloadResult]:
         """
         Download search results
@@ -827,6 +892,7 @@ class Audit(ServiceBase, AuditBase):
         Args:
             result_id: ID returned by the search API.
             format: Format for the records.
+            request_id: ID returned by the export API.
 
         Returns:
             URL where search results can be downloaded.
@@ -842,7 +908,10 @@ class Audit(ServiceBase, AuditBase):
             )
         """
 
-        input = DownloadRequest(result_id=result_id, format=format)
+        if request_id is None and result_id is None:
+            raise ValueError("must pass one of `request_id` or `result_id`")
+
+        input = DownloadRequest(request_id=request_id, result_id=result_id, format=format)
         return self.request.post("v1/download_results", DownloadResult, data=input.dict(exclude_none=True))
 
     def update_published_roots(self, result: SearchResultOutput):