+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support for Secure Audit Log's log stream API.
+- Support for Secure Audit Log's export API.
 
 ### Fixed